+++ CHANGES.rst
@@ -6,6 +6,9 @@ next release
 
 - OpenBGPD 6.1 support: enable large BGP communities support.
 - Improvement: the ``clients-from-peeringdb`` command now uses the `IX-F database <http://www.ix-f.net/ixp-database.html>`_ to show a list of IXP and their PeeringDB ID.
+- Improvement: enable NEXT_HOP rewriting for IPv6 blackhole filtering requests on OpenBGPD after `OpenBSD 6.1 fixup <https://github.com/openbsd/src/commit/f1385c8f4f9b9e193ff65d9f2039862d3e230a45>`_.
+
+  Related: `issue #3 <https://github.com/pierky/arouteserver/issues/3>`_.
 
 v0.5.0
 ------